+++ drake/examples/kuka_iiwa_arm/iiwa_common.cc
@@ -12,8 +12,8 @@
 #include "drake/common/text_logging.h"
 #include "drake/common/trajectories/piecewise_polynomial_trajectory.h"
 #include "drake/multibody/ik_options.h"
-#include "drake/multibody/kinematic_constraint/rigid_body_constraint.h"
 #include "drake/multibody/parsers/urdf_parser.h"
+#include "drake/multibody/rigid_body_constraint.h"
 #include "drake/multibody/rigid_body_ik.h"
 #include "drake/multibody/rigid_body_tree.h"
 #include "drake/util/drakeGeometryUtil.h"
