+++ drake/manipulation/planner/constraint_relaxing_ik.cc
@@ -5,8 +5,8 @@
 #include "drake/common/text_logging.h"
 #include "drake/multibody/ik_options.h"
 #include "drake/multibody/joints/floating_base_types.h"
-#include "drake/multibody/kinematic_constraint/rigid_body_constraint.h"
 #include "drake/multibody/parsers/urdf_parser.h"
+#include "drake/multibody/rigid_body_constraint.h"
 #include "drake/multibody/rigid_body_ik.h"
 
 namespace drake {
