+++ drake/multibody/BUILD
@@ -83,8 +83,8 @@ drake_cc_library(
 
 drake_cc_library(
     name = "rigid_body_constraint",
-    srcs = ["kinematic_constraint/rigid_body_constraint.cc"],
-    hdrs = ["kinematic_constraint/rigid_body_constraint.h"],
+    srcs = ["rigid_body_constraint.cc"],
+    hdrs = ["rigid_body_constraint.h"],
     deps = [
         ":rigid_body_tree",
     ],
@@ -397,7 +397,7 @@ drake_cc_googletest(
 
 drake_cc_googletest(
     name = "rigid_body_constraint_test",
-    srcs = ["kinematic_constraint/test/rigid_body_constraint_test.cc"],
+    srcs = ["test/rigid_body_constraint_test.cc"],
     data = ["//drake/examples/atlas:models"],
     deps = [
         ":rigid_body_constraint",
@@ -625,9 +625,6 @@ filegroup(
 # The extra_srcs are required here because add_lint_tests() doesn't understand
 # how to extract labels from select() functions yet.
 add_lint_tests(
-    cpplint_data = [
-        "kinematic_constraint/CPPLINT.cfg",
-    ],
     cpplint_extra_srcs = [
         "approximate_ik.cc",
         "test/test_approximate_ik.cc",
