+++ drake/multibody/approximate_ik.cc
@@ -5,7 +5,7 @@
 #include <gurobi_c++.h>
 
 #include "drake/multibody/ik_options.h"
-#include "drake/multibody/kinematic_constraint/rigid_body_constraint.h"
+#include "drake/multibody/rigid_body_constraint.h"
 #include "drake/multibody/rigid_body_ik.h"
 #include "drake/multibody/rigid_body_tree.h"
