+++ drake/multibody/dev/test/global_inverse_kinematics_test_util.cc
@@ -5,8 +5,8 @@
 
 #include "drake/common/eigen_matrix_compare.h"
 #include "drake/common/find_resource.h"
-#include "drake/multibody/kinematic_constraint/rigid_body_constraint.h"
 #include "drake/multibody/parsers/urdf_parser.h"
+#include "drake/multibody/rigid_body_constraint.h"
 #include "drake/multibody/rigid_body_ik.h"
 #include "drake/multibody/rigid_body_tree_construction.h"
