+++ drake/multibody/inverse_kinematics_backend.cc
@@ -15,7 +15,7 @@
 #include "drake/math/gradient.h"
 #include "drake/multibody/constraint_wrappers.h"
 #include "drake/multibody/ik_options.h"
-#include "drake/multibody/kinematic_constraint/rigid_body_constraint.h"
+#include "drake/multibody/rigid_body_constraint.h"
 #include "drake/multibody/rigid_body_tree.h"
 #include "drake/solvers/mathematical_program.h"
 #include "drake/solvers/snopt_solver.h"
