+++ drake/multibody/rigid_body_constraint.cc
@@ -1,4 +1,4 @@
-#include "drake/multibody/kinematic_constraint/rigid_body_constraint.h"
+#include "drake/multibody/rigid_body_constraint.h"
 
 // TODO(#2274) NOTE This file has so many cpplint errors that we have
 // whitelisted it in its entirety.  When the file is next rewritten or updates,
