+++ drake/multibody/rigid_body_constraint.h
