+++ drake/multibody/test/CMakeLists.txt
@@ -67,3 +67,10 @@ target_link_libraries(drakeRigidBodyLoopCompareToClone
   drakeRigidBodyFrameCompareToClone)
 
 add_subdirectory(rigid_body_tree)
+
+# Rigid body constraint unit tests
+drake_add_cc_test(NAME rigid_body_constraint_test)
+target_link_libraries(rigid_body_constraint_test
+    drakeMultibodyParsers
+    drakeRBM
+    drakeRigidBodyConstraint)
