+++ drake/multibody/test/rigid_body_constraint_test.cc
@@ -1,4 +1,4 @@
-#include "drake/multibody/kinematic_constraint/rigid_body_constraint.h"
+#include "drake/multibody/rigid_body_constraint.h"
 
 #include <memory>
