+++ drake/multibody/test/test_ik.cc
@@ -10,8 +10,8 @@
 #include "drake/common/find_resource.h"
 #include "drake/multibody/ik_options.h"
 #include "drake/multibody/joints/floating_base_types.h"
-#include "drake/multibody/kinematic_constraint/rigid_body_constraint.h"
 #include "drake/multibody/parsers/urdf_parser.h"
+#include "drake/multibody/rigid_body_constraint.h"
 #include "drake/multibody/rigid_body_ik.h"
 #include "drake/multibody/rigid_body_tree.h"
